requirements.txt
flask
flask_cors
pywin32
pyautogui
requests
pyopenssl
